Reporting correct progress for REST uploads

+++ fluxc/src/main/java/org/wordpress/android/fluxc/network/rest/wpcom/media/RestUploadRequestBody.java
@@ -36,7 +36,7 @@ public class RestUploadRequestBody extends BaseUploadRequestBody {
 
     @Override
     protected float getProgress(long bytesWritten) {
-        return bytesWritten / contentLength();
+        return (float) bytesWritten / contentLength();
     }
 
     @Override
